Спринт 18. Разделил один фрагмент на 2 отдельных фрагмента, добавил viewModel без логики.
