+++ app/src/main/java/com/practicum/playlistmaker/medialibrary/presentation/PlaylistsViewModel.kt
@@ -1,0 +1,6 @@
+package com.practicum.playlistmaker.medialibrary.presentation
+
+import androidx.lifecycle.ViewModel
+
+class PlaylistsViewModel(): ViewModel() {
+}
