+++ app/src/main/res/layout/fragment_favourite_tracks.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ImageView
+        android:id="@+id/icon_nothing_found"
+        android:contentDescription="@string/icon_nothing_found"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginTop="@dimen/dp_120"
+        style="@style/IconNothingFound"/>
+
+    <TextView
+        android:id="@+id/placeholderMessage"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/dp_20"
+        android:textSize="@dimen/text_19"
+        android:textColor="?attr/text_color"
+        android:fontFamily="@font/ys_display_medium"
+        android:gravity="center"
+        app:layout_constraintTop_toBottomOf="@id/icon_nothing_found"
+        />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
