+++ app/src/main/res/values/dimens.xml
@@ -35,6 +35,7 @@
     <dimen name="dp_61">61dp</dimen>
     <dimen name="dp_82">82dp</dimen>
     <dimen name="dp_110">110dp</dimen>
+    <dimen name="dp_120">120dp</dimen>
     <dimen name="dp_134">134dp</dimen>
     <dimen name="dp_150">150dp</dimen>
     <dimen name="dp_235">235dp</dimen>
